refactor: markdown-it's heading_open renderer rule
